+++ src/Processor/CheckoutProcessor.php
@@ -28,6 +28,7 @@ if (!defined('_PS_VERSION_')) {
 }
 
 use Cart;
+use Invertus\SaferPay\Api\Enum\TransactionStatus;
 use Invertus\SaferPay\Config\SaferPayConfig;
 use Invertus\SaferPay\Core\Payment\DTO\CheckoutData;
 use Invertus\SaferPay\EntityBuilder\SaferPayOrderBuilder;
@@ -39,6 +40,7 @@ use Invertus\SaferPay\Service\SaferPayInitialize;
 use Order;
 use PrestaShopException;
 use SaferPayOrder;
+use Validate;
 
 class CheckoutProcessor
 {
@@ -67,24 +69,24 @@ class CheckoutProcessor
     }
 
     public function run(CheckoutData $data) {
-
         $cart = new Cart($data->getCartId());
 
         if (!$cart) {
             throw CouldNotProcessCheckout::failedToFindCart($data->getCartId());
         }
 
-        if ($data->getIsAuthorizedOrder()) {
-            $this->processAuthorizedOrder($data, $cart);
-            return '';
+        if (!$data->getCreateAfterAuthorization()) {
+            $this->processCreateOrder($cart, $data->getPaymentMethod());
         }
 
-        try {
-            if (!$data->getCreateAfterAuthorization()) {
-                $this->processCreateOrder($cart, $data->getPaymentMethod());
-            }
-        } catch (\Exception $exception) {
-            throw CouldNotProcessCheckout::failedToCreateOrder($data->getCartId());
+        $authorizedStates = [
+            TransactionStatus::AUTHORIZED,
+            TransactionStatus::CAPTURED,
+        ];
+
+        if (in_array($data->getOrderStatus(), $authorizedStates)) {
+            $this->processAuthorizedOrder($data, $cart);
+            return '';
         }
 
         try {
@@ -121,6 +123,11 @@ class CheckoutProcessor
      */
     private function processCreateOrder(Cart $cart, $paymentMethod)
     {
+        // Notify and return webhooks triggers together leading into order created previously
+        if ($cart->orderExists()) {
+            return;
+        }
+
         $customer = new \Customer($cart->id_customer);
 
         $this->module->validateOrder(
@@ -182,36 +189,37 @@ class CheckoutProcessor
     private function processAuthorizedOrder(CheckoutData $data, Cart $cart)
     {
         try {
+            $this->processCreateOrder($cart, $data->getPaymentMethod());
+            $order = $this->getOrder($cart->id);
             $saferPayOrder = new SaferPayOrder($this->saferPayOrderRepository->getIdByCartId($cart->id));
 
-            if (!$saferPayOrder->id_order) {
-                $this->processCreateOrder($cart, $data->getPaymentMethod());
-            }
-
-            if (method_exists('Order', 'getIdByCartId')) {
-                $orderId = Order::getIdByCartId($cart->id);
-                $order = new Order($orderId);
+            if ($data->getOrderStatus() === TransactionStatus::AUTHORIZED) {
+                $saferPayOrder->authorized = true;
+                $order->setCurrentState(_SAFERPAY_PAYMENT_AUTHORIZED_);
             } else {
-                // For PrestaShop 1.6 use the alternative method
-                $orderId = Order::getOrderByCartId($cart->id);
-                $order = new Order($orderId);
+                $saferPayOrder->captured = true;
+                $order->setCurrentState(_SAFERPAY_PAYMENT_COMPLETED_);
             }
 
             $saferPayOrder->id_order = $order->id;
-
-            if ($data->getOrderStatus() === 'AUTHORIZED') {
-                $order->setCurrentState(_SAFERPAY_PAYMENT_AUTHORIZED_);
-                $saferPayOrder->authorized = 1;
-            } elseif ($data->getOrderStatus() === 'CAPTURED') {
-                $order->setCurrentState(_SAFERPAY_PAYMENT_COMPLETED_);
-                $saferPayOrder->captured = 1;
-            }
-
             $saferPayOrder->update();
-            $order->update();
-            return;
         } catch (\Exception $exception) {
             throw CouldNotProcessCheckout::failedToCreateOrder($data->getCartId());
+        }
+    }
+
+    /**
+     * @param int $cartId
+     *
+     * @return Order
+     */
+    private function getOrder($cartId)
+    {
+        if (method_exists('Order', 'getIdByCartId')) {
+            return new Order(Order::getIdByCartId($cartId));
+        } else {
+            // For PrestaShop 1.6 use the alternative method
+            return new Order(Order::getOrderByCartId($cartId));
         }
     }
 }